runtime: fix brk_ signature for plan9
with uintptr, the check for < 0 will never succeed in mem_plan9.go's
sbrk() because the brk_ syscall returns -1 on failure. fixes the plan9/amd64 build.

this failed on plan9/amd64 because of the attempt to allocate 136GB in mallocinit(),
which failed. it was just by chance that on plan9/386 allocations never failed.

Reviewed-on: https://go-review.googlesource.com/1590

+++ src/runtime/os_plan9.go
@@ -18,7 +18,7 @@ func seek(fd int32, offset int64, whence int32) int64
 func exits(msg *byte)
 
 //go:noescape
-func brk_(addr unsafe.Pointer) uintptr
+func brk_(addr unsafe.Pointer) int32
 
 func sleep(ms int32) int32
